#36: Reduced memory footprint of location parsing process
Changes:
 - Use low level stream json parser (jackson)
 - Avoid loading the whole contents of the location file as a string
 - Fixed leaked memory after index was built

Baseline:
  - after reading file: 66 MB
  - after parsing: 130 MB
  - after parsing+GC: 129MB

After changes:
  - after initializing stream: 4 MB
  - after parsing: 14 MB
  - after parsing+GC: 10 MB

We still need more memory to display the selector in the UI and build the index. After tests the application showed to behave well with ~60 Mb of memory.
It still didn't crash with 50 MB but it caused the VM to incurr in a lot of garbage collection.